Chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
openai
sqlalchemy
pydantic
python-jose
mysql-connector-python
mysqlclient
cryptography
mysql.connector
